Webcam Opencv Project/requirements.txt
numpy~=1.21.2
streamlit~=0.87.0
keras~=2.4.3
opencv-contrib-python-headless
tensorflow-cpu
streamlit_webrtc
